requirements.txt
openai>=1.0.0
faker>=15.0.0
pandas>=1.5.0
python-docx>=0.8.11
rich>=13.0.0
asyncio
